[Transform] refactor transform functions that utilize composite aggs
