+++ x-pack/plugin/transform/src/main/java/org/elasticsearch/xpack/transform/transforms/latest/Latest.java
@@ -9,20 +9,11 @@ package org.elasticsearch.xpack.transform.transforms.latest;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.elasticsearch.ElasticsearchException;
-import org.elasticsearch.ElasticsearchStatusException;
 import org.elasticsearch.action.ActionListener;
-import org.elasticsearch.action.index.IndexRequest;
-import org.elasticsearch.action.search.SearchAction;
-import org.elasticsearch.action.search.SearchRequest;
-import org.elasticsearch.action.search.SearchResponse;
-import org.elasticsearch.action.support.IndicesOptions;
 import org.elasticsearch.client.Client;
-import org.elasticsearch.common.collect.Tuple;
 import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.search.aggregations.AggregationBuilders;
-import org.elasticsearch.search.aggregations.Aggregations;
 import org.elasticsearch.search.aggregations.bucket.composite.CompositeAggregation;
 import org.elasticsearch.search.aggregations.bucket.composite.CompositeAggregationBuilder;
 import org.elasticsearch.search.aggregations.bucket.composite.CompositeValuesSourceBuilder;
@@ -30,15 +21,12 @@ import org.elasticsearch.search.aggregations.bucket.composite.TermsValuesSourceB
 import org.elasticsearch.search.aggregations.metrics.TopHits;
 import org.elasticsearch.search.aggregations.metrics.TopHitsAggregationBuilder;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
-import org.elasticsearch.xpack.core.ClientHelper;
-import org.elasticsearch.xpack.core.ml.utils.ExceptionsHelper;
 import org.elasticsearch.xpack.core.transform.TransformField;
 import org.elasticsearch.xpack.core.transform.transforms.SourceConfig;
 import org.elasticsearch.xpack.core.transform.transforms.TransformIndexerStats;
-import org.elasticsearch.xpack.core.transform.transforms.TransformProgress;
 import org.elasticsearch.xpack.core.transform.transforms.latest.LatestConfig;
-import org.elasticsearch.xpack.transform.transforms.Function;
 import org.elasticsearch.xpack.transform.transforms.IDGenerator;
+import org.elasticsearch.xpack.transform.transforms.common.AbstractCompositeAggFunction;
 import org.elasticsearch.xpack.transform.transforms.common.DocumentConversionUtils;
 
 import java.util.List;
@@ -48,23 +36,21 @@ import java.util.stream.Stream;
 import static java.util.Collections.emptyMap;
 import static java.util.stream.Collectors.toList;
 
-public class Latest implements Function {
+/**
+ * The latest transform function. This continually searches and processes results according to the passed {@link LatestConfig}
+ */
+public class Latest extends AbstractCompositeAggFunction {
 
     public static final int DEFAULT_INITIAL_MAX_PAGE_SEARCH_SIZE = 5000;
-    public static final int TEST_QUERY_PAGE_SIZE = 50;
 
-    private static final String COMPOSITE_AGGREGATION_NAME = "_transform";
     private static final String TOP_HITS_AGGREGATION_NAME = "_top_hits";
     private static final Logger logger = LogManager.getLogger(Latest.class);
 
     private final LatestConfig config;
 
-    // objects for re-using
-    private final CompositeAggregationBuilder cachedCompositeAggregation;
-
     public Latest(LatestConfig config) {
+        super(createCompositeAggregation(config));
         this.config = config;
-        this.cachedCompositeAggregation = createCompositeAggregation(config);
     }
 
     private static CompositeAggregationBuilder createCompositeAggregation(LatestConfig config) {
@@ -82,23 +68,6 @@ public class Latest implements Function {
     @Override
     public int getInitialPageSize() {
         return DEFAULT_INITIAL_MAX_PAGE_SEARCH_SIZE;
-    }
-
-    private SearchRequest buildSearchRequest(SourceConfig sourceConfig, int pageSize) {
-        SearchSourceBuilder sourceBuilder = new SearchSourceBuilder()
-            .query(sourceConfig.getQueryConfig().getQuery())
-            .runtimeMappings(sourceConfig.getRuntimeMappings());
-        buildSearchQuery(sourceBuilder, null, pageSize);
-        return new SearchRequest(sourceConfig.getIndex())
-            .source(sourceBuilder)
-            .indicesOptions(IndicesOptions.LENIENT_EXPAND_OPEN);
-    }
-
-    @Override
-    public SearchSourceBuilder buildSearchQuery(SearchSourceBuilder builder, Map<String, Object> position, int pageSize) {
-        cachedCompositeAggregation.aggregateAfter(position);
-        cachedCompositeAggregation.size(pageSize);
-        return builder.size(0).aggregation(cachedCompositeAggregation);
     }
 
     @Override
@@ -129,61 +98,6 @@ public class Latest implements Function {
     }
 
     @Override
-    public Tuple<Stream<IndexRequest>, Map<String, Object>> processSearchResponse(
-        SearchResponse searchResponse,
-        String destinationIndex,
-        String destinationPipeline,
-        Map<String, String> fieldTypeMap,
-        TransformIndexerStats stats
-    ) {
-        Aggregations aggregations = searchResponse.getAggregations();
-
-        // Treat this as a "we reached the end".
-        // This should only happen when all underlying indices have gone away. Consequently, there is no more data to read.
-        if (aggregations == null) {
-            return null;
-        }
-
-        CompositeAggregation compositeAgg = aggregations.get(COMPOSITE_AGGREGATION_NAME);
-        if (compositeAgg == null || compositeAgg.getBuckets().isEmpty()) {
-            return null;
-        }
-
-        Stream<IndexRequest> indexRequestStream =
-            compositeAgg.getBuckets().stream()
-                .map(bucket -> convertBucketToDocument(bucket, config, stats))
-                .map(document -> DocumentConversionUtils.convertDocumentToIndexRequest(document, destinationIndex, destinationPipeline));
-        return Tuple.tuple(indexRequestStream, compositeAgg.afterKey());
-    }
-
-    @Override
-    public void validateQuery(Client client, SourceConfig sourceConfig, ActionListener<Boolean> listener) {
-        SearchRequest searchRequest = buildSearchRequest(sourceConfig, TEST_QUERY_PAGE_SIZE);
-        client.execute(SearchAction.INSTANCE, searchRequest, ActionListener.wrap(response -> {
-            if (response == null) {
-                listener.onFailure(
-                    new ElasticsearchStatusException("Unexpected null response from test query", RestStatus.SERVICE_UNAVAILABLE)
-                );
-                return;
-            }
-            if (response.status() != RestStatus.OK) {
-                listener.onFailure(
-                    new ElasticsearchStatusException(
-                        "Unexpected status from response of test query: {}", response.status(), response.status())
-                );
-                return;
-            }
-            listener.onResponse(true);
-        }, e -> {
-            Throwable unwrapped = ExceptionsHelper.unwrapCause(e);
-            RestStatus status = unwrapped instanceof ElasticsearchException
-                ? ((ElasticsearchException) unwrapped).status()
-                : RestStatus.SERVICE_UNAVAILABLE;
-            listener.onFailure(new ElasticsearchStatusException("Failed to test query", status, unwrapped));
-        }));
-    }
-
-    @Override
     public void validateConfig(ActionListener<Boolean> listener) {
         listener.onResponse(true);
     }
@@ -194,44 +108,6 @@ public class Latest implements Function {
     }
 
     @Override
-    public void preview(
-        Client client,
-        Map<String, String> headers,
-        SourceConfig sourceConfig,
-        Map<String, String> mappings,
-        int numberOfBuckets,
-        ActionListener<List<Map<String, Object>>> listener
-    ) {
-        SearchRequest searchRequest = buildSearchRequest(sourceConfig, numberOfBuckets);
-        ClientHelper.assertNoAuthorizationHeader(headers);
-        ClientHelper.executeWithHeadersAsync(
-            headers,
-            ClientHelper.TRANSFORM_ORIGIN,
-            client,
-            SearchAction.INSTANCE,
-            searchRequest,
-            ActionListener.wrap(r -> {
-                Aggregations aggregations = r.getAggregations();
-                if (aggregations == null) {
-                    listener.onFailure(
-                        new ElasticsearchStatusException("Source indices have been deleted or closed.", RestStatus.BAD_REQUEST));
-                    return;
-                }
-                CompositeAggregation compositeAgg = aggregations.get(COMPOSITE_AGGREGATION_NAME);
-                TransformIndexerStats stats = new TransformIndexerStats();
-
-                List<Map<String, Object>> docs =
-                    compositeAgg.getBuckets().stream()
-                        .map(bucket -> convertBucketToDocument(bucket, config, stats))
-                        .map(DocumentConversionUtils::removeInternalFields)
-                        .collect(toList());
-
-                listener.onResponse(docs);
-            }, listener::onFailure)
-        );
-    }
-
-    @Override
     public SearchSourceBuilder buildSearchQueryForInitialProgress(SearchSourceBuilder searchSourceBuilder) {
         BoolQueryBuilder existsClauses = QueryBuilders.boolQuery();
         config.getUniqueKey().forEach(field -> existsClauses.must(QueryBuilders.existsQuery(field)));
@@ -239,7 +115,13 @@ public class Latest implements Function {
     }
 
     @Override
-    public void getInitialProgressFromResponse(SearchResponse response, ActionListener<TransformProgress> progressListener) {
-        progressListener.onResponse(new TransformProgress(response.getHits().getTotalHits().value, 0L, 0L));
+    protected Stream<Map<String, Object>> extractResults(
+        CompositeAggregation agg,
+        Map<String, String> fieldTypeMap,
+        TransformIndexerStats transformIndexerStats
+    ) {
+        return agg.getBuckets().stream()
+            .map(bucket -> convertBucketToDocument(bucket, config, transformIndexerStats))
+            .map(DocumentConversionUtils::removeInternalFields);
     }
 }
